Add alternate reboot detection
Normal detection checks the boot timestamp. While this method is very
reliable, it may not work if the device cannot track time and resets
to a certain epoch on every boot(e.g. devices with no network, no RTC,
no power).
Alternate detection checks two things: an intent fired with extras, and
a hint in the SharedPreferences. Both are triggered from the
ACTION_BOOT_COMPLETED BroadcastReceiver, so this method should be as
reliable as the OS can handle. Unfortunately that's not very much...